refactor: spread code
